book verification
